+++ packages/ckeditor5-dev-utils/tests/translations/cleanpofilecontent.js
@@ -11,7 +11,7 @@ const cleanPoFileContent = require( '../../lib/translations/cleanpofilecontent' 
 
 describe( 'translations', () => {
 	describe( 'cleanPoFileContent()', () => {
-		it( 'clean po files from personal data and add the special header', () => {
+		it( 'cleans po files from personal data and add the special header', () => {
 			const poFileContent =
 // eslint-disable-next-line max-len
 `# Copyright (c) 2003-2021, CKSource - Frederico Knabben. All rights reserved.
@@ -55,6 +55,89 @@ msgstr "URL del enllaz"
 `;
 
 			const result = cleanPoFileContent( poFileContent );
+
+			expect( result ).to.equal( expectedResult );
+		} );
+
+		it( 'does not add the copyright line if the source file misses it', () => {
+			const poFileContent =
+// eslint-disable-next-line max-len
+`# Translators:
+# Xuxxx Satxxx <xxxx@gmail.com>, 2017
+msgid ""
+msgstr ""
+"Last-Translator: Xuxxx Satxxx <xxxx@gmail.com>, 2017\\n"
+"Language: ast\\n"
+"Language-Team: Asturian (https://www.transifex.com/ckeditor/teams/11143/ast/)\\n"
+"Plural-Forms: nplurals=2; plural=(n != 1);\\n"
+
+msgctxt "Label for the url input in the Link dialog."
+msgid "Link URL"
+msgstr "URL del enllaz"
+`;
+
+			const expectedResult =
+// eslint-disable-next-line max-len
+`#                                     !!! IMPORTANT !!!
+#
+#         Before you edit this file, please keep in mind that contributing to the project
+#                translations is possible ONLY via the Transifex online service.
+#
+#         To submit your translations, visit https://www.transifex.com/ckeditor/ckeditor5.
+#
+#                   To learn more, check out the official contributor's guide:
+#     https://ckeditor.com/docs/ckeditor5/latest/framework/guides/contributing/contributing.html
+#
+msgid ""
+msgstr ""
+"Language: ast\\n"
+"Language-Team: Asturian (https://www.transifex.com/ckeditor/teams/11143/ast/)\\n"
+"Plural-Forms: nplurals=2; plural=(n != 1);\\n"
+
+msgctxt "Label for the url input in the Link dialog."
+msgid "Link URL"
+msgstr "URL del enllaz"
+`;
+
+			const result = cleanPoFileContent( poFileContent );
+
+			expect( result ).to.equal( expectedResult );
+		} );
+
+		it( 'removes the contribute url when passing options.simplifyLicenseHeader=true ', () => {
+			const poFileContent =
+`# Translators:
+# Xuxxx Satxxx <xxxx@gmail.com>, 2017
+msgid ""
+msgstr ""
+"Last-Translator: Xuxxx Satxxx <xxxx@gmail.com>, 2017\\n"
+"Language: ast\\n"
+"Language-Team: Asturian (https://www.transifex.com/ckeditor/teams/11143/ast/)\\n"
+"Plural-Forms: nplurals=2; plural=(n != 1);\\n"
+
+msgctxt "Label for the url input in the Link dialog."
+msgid "Link URL"
+msgstr "URL del enllaz"
+`;
+
+			const expectedResult =
+`#                                     !!! IMPORTANT !!!
+#
+#         Before you edit this file, please keep in mind that contributing to the project
+#                translations is possible ONLY via the Transifex online service.
+#
+msgid ""
+msgstr ""
+"Language: ast\\n"
+"Language-Team: Asturian (https://www.transifex.com/ckeditor/teams/11143/ast/)\\n"
+"Plural-Forms: nplurals=2; plural=(n != 1);\\n"
+
+msgctxt "Label for the url input in the Link dialog."
+msgid "Link URL"
+msgstr "URL del enllaz"
+`;
+
+			const result = cleanPoFileContent( poFileContent, { simplifyLicenseHeader: true } );
 
 			expect( result ).to.equal( expectedResult );
 		} );